+++ code/Tutorial_08/shaders/withTexture.vert
@@ -6,10 +6,10 @@
 layout(location = 0) in vec3 position;   // input:  attribute with index '0' with 3 elements per vertex
 layout(location = 1) in vec3 color;      // input:  attribute with index '1' with 3 elements (=rgb) per vertex
 layout(location = 2) in vec2 texcoords;  // input:  attribute with index '2' with 2 elements per vertex
-layout(location = 3) in int texnr;       // input:  attribute with index '3' - uint per vertex
+layout(location = 3) in float texnr;       // input:  attribute with index '3' - uint per vertex
 out vec4 fragColor;                      // output: computed fragmentation color
 out vec2 texCoord;                       // output: computed texture coordinates
-flat out int texID;                      // output: texture ID - mind the 'flat' attribute!
+flat out float texID;                      // output: texture ID - mind the 'flat' attribute!
 uniform mat4 worldToView;            // parameter: the camera matrix
 
 void main() {
